requirements.txt
cat > requirements.txt << 'EOF'
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
google-cloud-pubsub==2.18.4
gunicorn==21.2.0
EOF